+++ src/cpp/environment/border_wall.hpp
@@ -12,7 +12,7 @@ class BorderWall : public Silice3D::MeshObject, public Explodable {
 
  private:
   void ShowYouWonScreen();
-  virtual void ReactToExplosion(const glm::vec3& exp_position, float exp_radius) override;
+  virtual void ReactToExplosion(const glm::dvec3& exp_position, double exp_radius) override;
 };
 
 #endif
